refactor(p3): consolidate type directories into src/types/
Remove the parallel src/type/ (singular) directory and the root-level
src/types.ts by merging all their types into src/types/:

- src/type/levetypes.ts → src/types/leave.ts (added 'processing' status
  and managerRemarks field that levetypes had but leave.ts lacked)
- src/type/Hrtypes.ts → src/types/policy.ts (Policy was identical)
- src/type/Holiday.ts → src/types/holiday.ts (Holiday was identical)
- src/type/employee.ts + src/type/index.ts → unused, deleted
- src/types.ts → types merged into src/types/employee.ts (DepartmentFormData/
  Errors, richer Department fields) and new src/types/tenant.ts

After deletion of src/types.ts, existing '../../types' imports automatically
resolve to src/types/index.ts where the types are now exported from.
