--- src/type/Holiday.ts
@@ -1,6 +1,0 @@
-export interface Holiday {
-  id: string;
-  date: string; // format: "YYYY-MM-DD"
-  title: string;
-  description: string;
-}
