--- src/type/Hrtypes.ts
@@ -1,8 +1,0 @@
-// types.ts
-export interface Policy {
-  id: string;
-  name: string;
-  description: string;
-  type: string;
-  effectiveDate: string;
-}
