--- src/types.ts
@@ -1,36 +1,0 @@
-export interface Department {
-  id: string;
-  name: string;
-  nameAr: string;
-  description?: string;
-  descriptionAr?: string;
-  subtitle?: string;
-  subtitleAr?: string;
-}
-
-export interface DepartmentFormData {
-  name: string;
-  description?: string;
-}
-
-export interface DepartmentFormErrors {
-  name?: string;
-  description?: string;
-}
-
-export interface Tenant {
-  id: string;
-  name: string;
-  nameAr?: string;
-}
-
-export interface TenantFormData {
-  name: string;
-}
-
-export interface Company {
-  id: string;
-  name: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
