+++ src/types/employee.ts
@@ -2,6 +2,20 @@ export interface Department {
   id: string;
   name: string;
   nameAr: string;
+  description?: string;
+  descriptionAr?: string;
+  subtitle?: string;
+  subtitleAr?: string;
+}
+
+export interface DepartmentFormData {
+  name: string;
+  description?: string;
+}
+
+export interface DepartmentFormErrors {
+  name?: string;
+  description?: string;
 }
 
 export interface Designation {
