+++ src/types/tenant.ts
@@ -1,0 +1,16 @@
+export interface Tenant {
+  id: string;
+  name: string;
+  nameAr?: string;
+}
+
+export interface TenantFormData {
+  name: string;
+}
+
+export interface Company {
+  id: string;
+  name: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
